+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.19]
+- Field<T> should be accessed via factory method:
+  - Field.asInt()
+  - Field.as\<T>()
+
 ## [0.0.18]
 - small fix